// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleGroups_GetPacsDeviceAccessScheduleGroupForEditOutput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleGroups_PacsDeviceAccessScheduleGroupEditDto } from './FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleGroups_PacsDeviceAccessScheduleGroupEditDto';
import type { FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleItems_Dtos_PacsDeviceAccessScheduleItemEditDto } from './FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleItems_Dtos_PacsDeviceAccessScheduleItemEditDto';
export type FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleGroups_GetPacsDeviceAccessScheduleGroupForEditOutput = {
    deviceAccessScheduleGroup: FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleGroups_PacsDeviceAccessScheduleGroupEditDto;
    items?: Array<FireBytes_Unified_Wave_Pacs_PacsDeviceAccessScheduleItems_Dtos_PacsDeviceAccessScheduleItemEditDto> | null;
};